requirements.txt
Django==5.1.4
django-model-utils==5.0.0
requests==2.32.3
Celery==5.4.0
pymediainfo==6.1.0
fffw==7.0.0
mypy==1.10.1
types-requests==2.32.0.20240622
django-stubs==5.0.2
billiard==4.2.0
kombu==5.3.7
importlib-metadata==8.0.0
